requirements.txt
httpx>=0.27,<1.0
python-dotenv>=1.0,<2.0
rich>=13.7,<15
mcp>=1.1,<2.0
pytest>=7.0
pytest-asyncio>=0.23
